(refactor) : Results, Result, CarState에서 의존성을 낮추기 위한 리팩토링
