+++ src/main/java/racingcar/domain/Result.java
@@ -3,6 +3,7 @@ package racingcar.domain;
 import java.util.Collections;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.stream.Collectors;
 
 public class Result {
 
@@ -20,10 +21,21 @@ public class Result {
         return Collections.unmodifiableList(result);
     }
 
-    public int getMaxPosition() {
+    public List<String> getWinners() {
+        return result.stream()
+            .filter(carState -> isMaxPosition(carState.getPosition()))
+            .map(CarState::getName)
+            .collect(Collectors.toList());
+    }
+
+    private int getMaxPosition() {
         return result.stream()
             .mapToInt(CarState::getPosition)
             .max()
             .orElseThrow(NoSuchElementException::new);
+    }
+
+    private boolean isMaxPosition(int position) {
+        return getMaxPosition() == position;
     }
 }
